refactor: refactor import module and module
